Move source files and json files to their own folders
