+++ JSON/rentals.json
@@ -1,0 +1,1 @@
+[]
