--- people.json
@@ -1,1 +1,0 @@
-[{"json_class":"Student","a":["Sdf",0,"432",{"parent_permission":false}]}]
