--- rentals.json
@@ -1,1 +1,0 @@
-[{"json_class":"Rental","a":["654",{"json_class":"Book","a":["gi","hud"]},{"json_class":"Student","a":["Sdf",0,"432",{"parent_permission":false}]}]}]
